+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unreleased
 
 - feat: Store SentryEnvelopes in extra path #468
+- feat: Adds setUser to SentrySDK and SentryHub #467
 - feat: Add auto session starting for macOS #463
 
 ## 5.0.0-beta.6